Started blockout
I got the vague shape of the ground, the tower in the middle, and the natural wall separating the two sides. It's not much, but should be enough for everyone to be able to start on their own section.

If you're going to use the landscaping tool for your area, be sure to mess with the LOD0 settings of the object so your land doesn't start moving as you walk around on it